Remove unneeded elements from SignIn page
