+++ client/src/App.js
@@ -1,23 +1,14 @@
 import './App.css'
 import React from 'react'
-import BasicButtonGroup from './components/BasicButtonGroup/BasicButtonGroup'
-import SimpleMenu from './components/SimpleMenu/SimpleMenu'
-import SignUp from './components/SignUp/SignUp'
-// import Box from '@material-ui/core/Box';
-// import SimpleContainer from './components/Container/Container';
+import SignIn from './pages/SignIn/SignIn'
+
 
 function App() {
   return (
     <div className='App'>
-      {/* <Box component="span" color="primary" m={3}> */}
-      {/* <SimpleContainer> */}
       <header className='App-header'>
-        <SignUp />
-        <SimpleMenu />
-        <BasicButtonGroup />
+        <SignIn />
         </header>
-      {/* </SimpleContainer> */}
-      {/* </Box> */}
     </div>
   );
 }
