--- client/src/components/BasicButtonGroup/BasicButtonGroup.jsx
@@ -1,30 +1,0 @@
-import React from 'react';
-import Button from '@material-ui/core/Button';
-import ButtonGroup from '@material-ui/core/ButtonGroup';
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    '& > *': {
-      margin: theme.spacing(),
-    },
-  },
-}));
-
-export default function BasicButtonGroup() {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.root}>
-      <ButtonGroup variant="contained" color="primary" aria-label="contained primary button group">
-        <Button>Beans</Button>
-        <Button>Cheese</Button>
-        <Button>Chicken</Button>
-      </ButtonGroup>
-    </div>
-  );
-}
-
