+++ client/src/components/Button/Button.jsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import Button from '@material-ui/core/Button';
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    '& > *': {
+      margin: theme.spacing(1),
+    },
+  },
+}));
+
+export default function SignInButton() {
+
+  const classes = useStyles();
+
+  return (
+    <div className={classes.root}>
+      <Button
+        type="submit"
+        fullWidth
+        variant="contained"
+        color="primary"
+        className={classes.submit}
+      >
+        Sign In
+      </Button>
+    </div>
+  );
+}
